requirements.txt
fastapi==0.128.8
uvicorn==0.39.0

python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==3.2.2

python-multipart==0.0.20